requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.11.0
beautifulsoup4==4.12.3
black==24.4.3.dev27+g7fa1faf
bleach==4.1.0
blinker==1.9.0
boto3==1.34.82
botocore==1.34.82
Brotli==1.0.9
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.2.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
eland==8.12.1
elastic-transport==8.15.1
elasticsearch==8.17.0
embed_anything==0.2.2
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.5
idna==3.7
ipykernel==6.28.0
ipython==8.25.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.10
keras==3.7.0
kiwisolver==1.4.7
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-elasticsearch==0.2.2
langchain-google-genai==1.0.8
langchain-huggingface==0.0.3
langchain-ollama==0.2.2
langchain-openai==0.1.21
langchain-text-splitters==0.3.4
langgraph==0.2.4
langgraph-checkpoint==1.0.3
langsmith==0.1.147
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.19.1
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
notebook==7.0.8
notebook_shim==0.2.3
numpy==1.26.4
ollama==0.4.4
openai==1.40.8
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.6
overrides==7.4.0
packaging==24.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pillow==10.4.0
pip==24.0
platformdirs==3.10.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.1.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pydantic-settings==2.7.1
pydeck==0.9.1
Pygments==2.15.1
pyparsing==3.1.2
pypdf==5.1.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.30.2
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.10.6
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.14.0
Send2Trash==1.8.2
sentence-transformers==3.0.1
setuptools==69.5.1
sip==6.7.12
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.2.0
streamlit==1.41.1
streamlit-keyup==0.2.4
streamlit-searchbox==0.1.20
sympy==1.13.1
tavily-python==0.5.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
terminado==0.17.1
tf_keras==2.18.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.2.1
tokenize-rt==5.2.0
tokenizers==0.19.1
toml==0.10.2
torch==2.4.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.5
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.43.0
widgetsnbextension==4.0.10
wrapt==1.17.0
xgboost==2.1.3
xxhash==3.5.0
yarl==1.9.4
